Remove full-assembly.?pp; the little remaining code is now in algebraic.?pp
