+++ fem/ceed/algebraic.hpp
@@ -26,6 +26,15 @@ namespace ceed
 {
 
 #ifdef MFEM_USE_CEED
+
+/** @brief Assembles a CeedOperator as an mfem::SparseMatrix
+
+    In parallel, this assembles independently on each processor, that is, it
+    assembles at the L-vector level. The assembly procedure is always
+    performed on the host, but this works for operators stored on device
+    by copying memory. */
+int CeedOperatorFullAssemble(CeedOperator op, SparseMatrix **mat);
+
 /** @brief A way to use algebraic levels in a Multigrid object
 
     This is analogous to a FiniteElementSpace but with no Mesh information,
